requirements.txt
requests
numpy
pandas
python-dotenv
scikit-learn
matplotlib
seaborn
google-generativeai
anthropic